refactor(ui): move all pages to new page layout component
