+++ code/workspaces/web-app/src/pages/LandingPage.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
-import Segment from '../components/app/Segment';
+import Page from './Page';
 
 const LandingPage = () => (
-  <Segment>
-    <Typography gutterBottom variant="h4">Home Page</Typography>
-    <p>This is the data lab home page.</p>
-  </Segment>
+  <Page title="Dashboard">
+    <Typography variant="body1">This is the data lab home page.</Typography>
+  </Page>
 );
 
 export default LandingPage;
