+++ code/workspaces/web-app/src/pages/SparkPage.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
-import Segment from '../components/app/Segment';
+import Page from './Page';
 
 const SparkPage = () => (
-  <Segment>
-    <Typography gutterBottom variant="h4">Spark</Typography>
-    <p>Apache Spark is an open-source cluster-computing framework.</p>
-  </Segment>
+  <Page title="Spark">
+    <Typography variant="body1">Apache Spark is an open-source cluster-computing framework.</Typography>
+  </Page>
 );
 
 export default SparkPage;
